test(dp): Remove DB references from integraion tests
